requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[dotenv]
pyjwt